Fixed|libappfw: Apply UI scale factor; non-integer numbers in VariableChoiceWidget

+++ doomsday/sdk/libappfw/src/widgets/variablechoicewidget.cpp
@@ -13,7 +13,7 @@
  * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
  * General Public License for more details. You should have received a copy of
  * the GNU Lesser General Public License along with this program; if not, see:
- * http://www.gnu.org/licenses</small> 
+ * http://www.gnu.org/licenses</small>
  */
 
 #include "de/VariableChoiceWidget"
@@ -58,7 +58,7 @@ DENG2_OBSERVES(Variable, Change  )
         if(!var) return;
 
         var->audienceForChange() -= this;
-        var->set(NumberValue(self.selectedItem().data().toInt()));
+        var->set(NumberValue(self.selectedItem().data().toDouble()));
         var->audienceForChange() += this;
     }
